Add files via upload
In this paper, I have propounded a system which is capable to detect
any trespass and intrusion as well as fire and can eliminate fire in
control manner provide the exact location of the affected region. IT is
also capable of detecting false alarm and stop it, it is also capable of
detecting any toxic gas leakage and take action to it.
Raspberry Pi 3 has been used for flow control which are integrated
with a couple of sensors and camera including flame sensor, gas
sensor, PIR sensor, MCP3008-ADC convertor, relay, electricity
operated motors, ESP32 802.11 camera, GSM module I have
provided a confirmation of the fire suspecting system to avoid any
false alarm. The system will immediately send an email along with
the image of the affected spot and device location. Also an emergency
message is sent. In case of any fire or gas leak or trespass. It may also
able to capture the face of the trespasser